Have an alternative way to display species with no data on MOL
