+++ src/components/species-search/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import SpeciesSearchComponent from './species-search-component';
+
+function SpeciesSearch(props) {
+  return <SpeciesSearchComponent {...props} />;
+}
+
+export default SpeciesSearch;
