+++ src/containers/sidebars/dashboard-sidebar/data-layers/data-layers-styles.module.scss
@@ -33,6 +33,10 @@
     color: var(--font-color);
   }
 
+  p{
+    color: var(--font-color);
+  }
+
   .data{
     display: flex;
     flex-direction: column;
