+++ src/containers/sidebars/dashboard-sidebar/species-home/species-home-styles.module.scss
@@ -37,88 +37,6 @@
       color: var(--font-color);
     }
 
-    .explore{
-      position: relative;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 2rem;
-      margin-bottom: 20px;
-
-      .search{
-        border: 1px solid var(--border-color);
-        padding: 5px;
-        border-radius: 8px;
-        flex-grow: 1;
-        background: #{$white};
-
-        input[type="text"]{
-          color: var(--border-color);
-          width: 100%;
-        }
-
-        ::placeholder{
-          color: var(--border-color);
-        }
-
-        svg{
-          fill: var(--border-color);
-          color: var(--border-color);
-
-          path{
-            fill: var(--border-color);
-            color: var(--border-color);
-          }
-        }
-
-        &.showResults{
-          border-bottom-left-radius: 0;
-          border-bottom-right-radius: 0;
-        }
-      }
-
-      .searchResults{
-        position: absolute;
-        top: 100%;
-        left: 5px;
-        background-color: var(--border-color);
-        width: 329px;
-        border-bottom-right-radius: 5px;
-        border-bottom-left-radius: 5px;
-        padding: 8px 10px 5px;
-        margin: -2px 0 0 -1px;
-        list-style: none;
-
-        li{
-          button{
-            cursor: pointer;
-            padding: 5px;
-            color: #{$white};
-            text-align: left;;
-
-            span{
-              font-size: $font-size-sm;
-              font-style: italic;
-            }
-
-            &:hover{
-              background-color:#{$white};
-              color: var(--border-color);
-            }
-          }
-        }
-      }
-
-      button{
-        width:auto;
-        color: var(--save-button-text);
-        background-color: var(--border-color);
-        &:hover {
-          background-color: var(--border-color-hover);
-        }
-      }
-    }
-
     .mostPopular{
       .species{
         display: flex;
